+++ Source/abci/Importer/AlembicImporter.cpp
@@ -51,6 +51,11 @@ abciAPI bool aiContextGetIsHDF5(aiContext* ctx)
     return ctx ? ctx->getIsHDF5() : false;
 }
 
+abciAPI const char* aiContextGetApplication(aiContext* ctx)
+{
+    return ctx ? ctx->getApplication() : "";
+}
+
 abciAPI void aiContextSetConfig(aiContext* ctx, const aiConfig* conf)
 {
     if (ctx)
@@ -188,7 +193,8 @@ abciAPI aiSample* aiSchemaGetSample(aiSchema * schema)
 
 abciAPI void aiSchemaUpdateSample(aiSchema* schema, const abcSampleSelector *ss)
 {
-    if (schema) {
+    if (schema)
+    {
         schema->markForceUpdate();
         schema->updateSample(*ss);
     }