+++ Source/abci/Importer/AlembicImporter.h
@@ -280,6 +280,7 @@ abciAPI aiContext*      aiContextCreate(int uid);
 abciAPI void            aiContextDestroy(aiContext* ctx);
 abciAPI bool            aiContextLoad(aiContext* ctx, const char *path);
 abciAPI bool            aiContextGetIsHDF5(aiContext* ctx);
+abciAPI const char*     aiContextGetApplication(aiContext* ctx);
 abciAPI void            aiContextSetConfig(aiContext* ctx, const aiConfig* conf);
 abciAPI int             aiContextGetTimeSamplingCount(aiContext* ctx);
 abciAPI aiTimeSampling* aiContextGetTimeSampling(aiContext* ctx, int i);