+++ com.unity.formats.alembic/Runtime/Scripts/Exporter/AlembicExporter_impl.cs
@@ -510,8 +510,8 @@ namespace UnityEngine.Formats.Alembic.Util
                 abcObject = abc;
             }
 
-            public override void Setup(Component c) {}
-            public override void Capture() {}
+            public override void Setup(Component c) { }
+            public override void Capture() { }
         }
 
         [CaptureTarget(typeof(Transform))]
@@ -635,7 +635,7 @@ namespace UnityEngine.Formats.Alembic.Util
                 }
                 else
                 {
-                    const float deg2rad =  Mathf.PI / 180;
+                    const float deg2rad = Mathf.PI / 180;
                     dst.focalLength = (float)(Screen.height / 2 / Math.Tan(deg2rad * src.fieldOfView / 2));
                     dst.sensorSize = new Vector2(Screen.width, Screen.height);
                 }
@@ -745,7 +745,7 @@ namespace UnityEngine.Formats.Alembic.Util
                     {
                         if (m_meshBake == null)
                         {
-                            m_meshBake = new Mesh {name = m_target.name};
+                            m_meshBake = new Mesh { name = m_target.name };
                         }
 
                         m_meshBake.Clear();
@@ -944,17 +944,6 @@ namespace UnityEngine.Formats.Alembic.Util
 
 #endif
 
-        T[] GetTargets<T>() where T : Component
-        {
-            if (m_settings.Scope == ExportScope.TargetBranch && TargetBranch != null)
-            {
-                return TargetBranch.GetComponentsInChildren<T>();
-            }
-            else
-            {
-                return GameObject.FindObjectsOfType<T>();
-            }
-        }
 
         Component[] GetTargets(Type type)
         {
@@ -1050,7 +1039,7 @@ namespace UnityEngine.Formats.Alembic.Util
                     var cr = m_capturerTable[node.componentType];
                     node.componentCapturer = Activator.CreateInstance(cr.type) as ComponentCapturer;
                     node.componentCapturer.recorder = this;
-                    node.componentCapturer.parent = node.transformCapturer;;
+                    node.componentCapturer.parent = node.transformCapturer; ;
                     node.componentCapturer.timeSamplingIndex = timeSamplingIndex;
                     node.componentCapturer.Setup(component);
                 }