+++ com.unity.formats.alembic/CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0-pre.3] - 2021-10-02
+### Added
+### Changed
+### Fixed
+
+## [2.3.0-pre.2] - 2021-09-02
+### Added
+### Changed
+### Fixed
+- Prevent mesh import failure when changing import options after having set up material assignment overrides.
+- Fixed a UI issue affecting the use of Importer options in multi-editing context.
+- Prevent import failure when material mapping data is inconsistent with Alembic node hierarchy.
+
 ## [2.3.0-pre.1] - 2021-07-11
 ### Added
 - Added support for importing the cage mesh of SubD Alembic nodes.
@@ -11,12 +24,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Material tab includes a tool to automatically assign materials based on the Face Set names of the original Alembic file.
 
 ### Changed
-- Alembic now exported GameObjects now inclue also GameObjects that change only their transform.
+- GameObjects that only have a Transform component are now included when exporting Alembic files.
 - Automatically set the timeline clip length when dragging an AlembicStreamPlayer onto a timeline Alembic track.
 - Unsupported schema nodes are imported as NO-OP GameObjects.
 ### Fixed
 - Fixed a bug that caused Unity Cloth Alembic export to ignore the world scale.
 - Fixed a bug that caused an error when manually adding an AlembicStreamPlayer Component.
+- Prevented material assignments from being lost for a streamed external Alembic file when changing the Alembic source file link.
 
 ## [2.2.0] - 2021-06-10
 ### Added